update(Router): New version 3.0.0-alpha.3, update packages, update installation documentation
