+++ package.json
@@ -42,9 +42,9 @@
     "path": "^0.12.7",
     "prismjs": "^1.4.1",
     "run-sequence": "^1.1.5",
-    "systemjs-builder": "^0.15.15",
-    "typescript": "^1.8.10",
-    "typings": "^1.0.4",
+    "systemjs-builder": "^0.15.20",
+    "typescript": "^1.9.0-dev.20160615-1.0",
+    "typings": "^1.0.5",
     "vinyl-buffer": "^1.0.0",
     "vinyl-paths": "^2.1.0",
     "vinyl-source-stream": "^1.1.0"
@@ -56,7 +56,7 @@
     "@angular/http": "2.0.0-rc.1",
     "@angular/platform-browser": "2.0.0-rc.1",
     "@angular/platform-browser-dynamic": "2.0.0-rc.1",
-    "@angular/router": "2.0.0-rc.1",
+    "@angular/router": "3.0.0-alpha.3",
     "bootstrap": "^4.0.0-alpha.2",
     "font-awesome": "^4.6.1",
     "rxjs": "5.0.0-beta.6",
@@ -69,15 +69,15 @@
     "@angular/http": "2.0.0-rc.1",
     "@angular/platform-browser": "2.0.0-rc.1",
     "@angular/platform-browser-dynamic": "2.0.0-rc.1",
-    "@angular/router": "2.0.0-rc.1",
+    "@angular/router": "3.0.0-alpha.3",
     "bootstrap": "^4.0.0-alpha.2",
     "core-js": "^2.2.1",
-    "es6-shim": "^0.35.0",
+    "es6-shim": "^0.35.1",
     "font-awesome": "^4.6.1",
     "hammerjs": "^2.0.8",
     "reflect-metadata": "^0.1.3",
     "rxjs": "5.0.0-beta.6",
-    "systemjs": "^0.19.25",
+    "systemjs": "^0.19.30",
     "zone.js": "0.6.12"
   }
 }
