+++ src/demo.ts
@@ -1,96 +1,10 @@
-import "node_modules/core-js/client/library.min.js";
 import {Component, ViewEncapsulation, provide, ChangeDetectionStrategy, enableProdMode} from "@angular/core";
 import {FORM_DIRECTIVES, FORM_PROVIDERS, CORE_DIRECTIVES, LocationStrategy, HashLocationStrategy } from "@angular/common";
 import {bootstrap} from "@angular/platform-browser-dynamic";
-import {Routes, Route, ROUTER_DIRECTIVES, ROUTER_PROVIDERS} from "@angular/router";
+import {ROUTER_DIRECTIVES} from '@angular/router';
 import {FUELUI_COMPONENT_PROVIDERS, FUELUI_DIRECTIVE_PROVIDERS, FUELUI_PIPE_PROVIDERS, CodeHighlighter} from "./fuel-ui";
-import {AccordionDemo, AlertDemo, CarouselDemo, CollapseDemo, DatePickerDemo, DateRangePickerDemo, DropdownDemo, 
-    InfiniteScrollerDemo, ModalDemo, PaginationDemo, ProgressDemo, TableSortableDemo, AnimationDemo, CodeHighlighterDemo,
-    TooltipDemo, FormatDemo, MapToIterableDemo, OrderByDemo, RangeDemo, SliderDemo, TabDemo, TagDemo, TimePickerDemo} from './fuel-ui-demo';
+import {APP_ROUTER_PROVIDERS} from './demo.routes';
 
-@Component({
-    template: `
-    <div class="jumbotron jumbotron-fluid">
-        <div class="container">
-            <h2 class="display-3">Fuel-UI</h2>
-            <p class="lead">Fuel-UI is a collection of native <a href="http://angular.io" target="_blank">Angular 2</a> components, directives, and pipes for <a href="http://v4-alpha.getbootstrap.com/" target="_blank">Bootstrap 4</a>.</p>
-            
-            <a href="https://github.com/FuelInteractive/fuel-ui/releases" target="_blank" class="btn btn-fuel">Download <i class="fa fa-download"></i></a> 
-            <a href="https://github.com/FuelInteractive/fuel-ui" target="_blank" class="btn btn-fuel">View on GitHub <i class="fa fa-external-link"></i></a> 
-            <a href="https://www.npmjs.com/package/fuel-ui" target="_blank" class="btn btn-fuel">View npm Package <i class="fa fa-external-link"></i></a>
-        </div>
-    </div>
-    
-    <p>Fuel-UI is developed by <a href="http://fueltravel.com" target="_blank">Fuel Travel</a>, a company with years of expertise in the travel marketing industry. For project news and updates, follow us on <a href="http://twitter.com/fueltravel" target="_blank">twitter</a>.</p>`
-})
-export class DemoHome {
-    
-}
-
-@Component({
-    template: `
-<div class="row">
-    <div class="col-md-12">
-        <div class="card card-block">
-            <h2 class="card-title">Installation</h2>
-            <p class="card-text">Fork our Quickstart! <a href="https://github.com/coryshaw1/ng2-play/" target="_blank">https://github.com/coryshaw1/ng2-play/</a></p>
-        </div>
-    </div>
-</div>
-
-<p>If you would like to add Fuel-UI to your Angular2 project through npm manually, do the following:</p>
-
-<p><code>npm install fuel-ui font-awesome bootstrap@^4.0.0-alpha.2 --save</code></p>
-            
-<p>Then simply add the proper script tags to your <code>index.html</code></p>
-            
-<pre>
-<code class="language-markup" code-highlight>
-&lt;head&gt;
-    &lt;link rel=&quot;stylesheet&quot; href=&quot;node_modules/font-awesome/css/font-awesome.min.css&quot; /&gt;
-    &lt;link rel=&quot;stylesheet&quot; href=&quot;node_modules/bootstrap/dist/css/bootstrap.min.css&quot; /&gt;
-    &lt;link rel=&quot;stylesheet&quot; href=&quot;node_modules/fuel-ui/bundles/fuel-ui.min.css&quot; /&gt;
-&lt;/head&gt;
-
-...
-
-&lt;!-- All your SystemJS, Angular2, Rx, etc. scripts first! --&gt;
-&lt;script src=&quot;node_modules/fuel-ui/bundles/fuel-ui.min.js&quot;&gt;&lt;/script&gt;
-</code>
-</pre>`,
-directives: [CodeHighlighter]
-})
-export class InstallationComponent {
-    
-}
-
-@Routes([
-    { path: '/', component: DemoHome},
-    { path: '/installation', component: InstallationComponent},
-    { path: '/component/accordion', component: AccordionDemo },
-    { path: '/component/alert', component: AlertDemo },
-    { path: '/component/carousel', component: CarouselDemo },
-    { path: '/component/datepicker', component: DatePickerDemo },
-    { path: '/component/daterangepicker', component: DateRangePickerDemo },
-    { path: '/component/dropdown', component: DropdownDemo },
-    { path: '/component/infinitescroller', component: InfiniteScrollerDemo },
-    { path: '/component/modal', component: ModalDemo },
-    { path: '/component/pagination', component: PaginationDemo },
-    { path: '/component/progress', component: ProgressDemo },
-    { path: '/component/slider', component: SliderDemo },
-    { path: '/component/tab', component: TabDemo },
-    { path: '/component/tablesortable', component: TableSortableDemo },
-    { path: '/component/tag', component: TagDemo },
-    { path: '/component/timepicker', component: TimePickerDemo },
-    { path: '/directive/animation', component: AnimationDemo },
-    { path: '/directive/codehighlighter', component: CodeHighlighterDemo },
-    { path: '/directive/collapse', component: CollapseDemo },
-    { path: '/directive/tooltip', component: TooltipDemo },
-    { path: '/pipe/format', component: FormatDemo },
-    { path: '/pipe/maptoiterable', component: MapToIterableDemo },
-    { path: '/pipe/orderby', component: OrderByDemo },
-    { path: '/pipe/range', component: RangeDemo}
-])
 @Component({
 	selector: "fuel-ui",
     template: `
@@ -130,6 +44,7 @@ export class InstallationComponent {
                         <li><a (click)="clickNavLink()" [routerLink]="['/component', 'tab']">Tabs</a></li>
                         <li><a (click)="clickNavLink()" [routerLink]="['/component', 'tablesortable']">TableSortable</a></li>
                         <li><a (click)="clickNavLink()" [routerLink]="['/component', 'tag']">Tags</a></li>
+                        <li><a (click)="clickNavLink()" [routerLink]="['/component', 'textexpander']">TextExpander</a></li>
                         <li><a (click)="clickNavLink()" [routerLink]="['/component', 'timepicker']">TimePicker</a></li>
                     </accordion-item>
                     <accordion-item #directiveNav [open]="false">
@@ -192,8 +107,8 @@ export class DemoComponent {
 // enableProdMode();
 
 bootstrap(DemoComponent, [
-    ROUTER_PROVIDERS,
+    APP_ROUTER_PROVIDERS,
     FORM_PROVIDERS,
     provide(LocationStrategy, { useClass: HashLocationStrategy }),
     FUELUI_COMPONENT_PROVIDERS
-]);
+]).catch(err => console.error(err));
